+++ CHANGELOG.md
@@ -32,6 +32,7 @@ CHANGELOG
     - support for zip
     - new parameters `try_zip`, `try_gzip` and `try_tar` to control which compressions are tried.
     - rewritten to an iterative approach
+  - add `file_name_from_response` to extract a file name from a Response object for downloaded files.
 
 ### Development